backend/requirements.txt
Pillow>=10.0
pillow-heif
Flask
Flask-CORS
werkzeug>=2.3.4
gunicorn
Flask-SQLAlchemy
SQLAlchemy
psycopg2-binary
PyJWT
stripe
